# api/api/openapi.yaml
openapi: 3.0.3
info:
  title: EHR API
  description: |
    Headless Electronic Health Record API with dual interfaces:
    - FHIR R4-compliant endpoints under /fhir
    - Operational REST endpoints under /api/v1
  version: 0.1.0
  contact:
    name: EHR Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /fhir/Coverage:
    get:
      tags: [Coverage]
      summary: Search coverages (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - $ref: '#/components/parameters/FHIRStatus'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Coverage]
      summary: Create coverage (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Coverage'
      responses:
        '201':
          description: Coverage created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Coverage'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /fhir/Claim:
    get:
      tags: [Claim]
      summary: Search claims (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - $ref: '#/components/parameters/FHIRStatus'
        - name: use
          in: query
          schema:
            type: string
            enum: [claim, preauthorization, predetermination]
          description: Claim use
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Claim]
      summary: Create claim (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Claim'
      responses:
        '201':
          description: Claim created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Claim'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/claims:
    get:
      tags: [Claim]
      summary: List claims
      parameters:
        - name: patient_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Paginated list of claims
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      tags: [Claim]
      summary: Create claim
      responses:
        '201':
          description: Claim created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/claims/{id}:
    get:
      tags: [Claim]
      summary: Get claim by ID
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Claim details
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Claim]
      summary: Update claim
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Claim updated
    delete:
      tags: [Claim]
      summary: Delete claim
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Claim deleted

  # ── Coverages (Operational) ────────────────────────────────────────
  /api/v1/coverages:
    get:
      tags: [Coverage]
      summary: List coverages
      parameters:
        - name: patient_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Paginated list of coverages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    post:
      tags: [Coverage]
      summary: Create coverage
      responses:
        '201':
          description: Coverage created
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/coverages/{id}:
    get:
      tags: [Coverage]
      summary: Get coverage by ID
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Coverage details
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Coverage]
      summary: Update coverage
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Coverage updated
    delete:
      tags: [Coverage]
      summary: Delete coverage
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '204':
          description: Coverage deleted

  # ── Practitioners (Operational) ────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. In development mode, the server accepts any
        well-formed JWT. In production, tokens are validated against
        the configured JWKS endpoint.
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Internal UUID
    FHIRPatient:
      name: patient
      in: query
      schema:
        type: string
      description: Patient reference (e.g. Patient/123)
    FHIRStatus:
      name: status
      in: query
      schema:
        type: string
      description: Resource status code
...
    Count:
      name: _count
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of results per page (FHIR)
    Offset:
      name: _offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Result offset for pagination (FHIR)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of results per page
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Result offset for pagination
  responses:
    FHIRBundle:
      description: FHIR Bundle (searchset)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/Bundle'
    BadRequest:
      description: Invalid request
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
...
    NotFound:
      description: Resource not found
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
    InternalError:
      description: Internal server error
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
  schemas:
...
    Bundle:
      type: object
      properties:
        resourceType:
          type: string
          enum: [Bundle]
        type:
          type: string
          enum: [searchset]
        total:
          type: integer
          description: Total number of matching resources
        link:
          type: array
          items:
            type: object
            properties:
              relation:
                type: string
                enum: [self, next, previous]
              url:
                type: string
                format: uri
        entry:
          type: array
          items:
            type: object
            properties:
              fullUrl:
                type: string
                format: uri
              resource:
                type: object
                description: The FHIR resource
              search:
                type: object
                properties:
                  mode:
                    type: string
                    enum: [match, include, outcome]
    OperationOutcome:
      type: object
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          items:
            type: object
            required: [severity, code]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
                description: Error code (not-found, invalid, etc.)
              diagnostics:
                type: string
                description: Human-readable diagnostic message
...
    Meta:
      type: object
      properties:
        versionId:
          type: string
        lastUpdated:
          type: string
          format: date-time
    Reference:
      type: object
      properties:
        reference:
          type: string
          description: Relative reference (e.g. Patient/123)
        display:
          type: string
    CodeableConcept:
      type: object
      properties:
        coding:
          type: array
          items:
            $ref: '#/components/schemas/Coding'
        text:
          type: string
    Coding:
      type: object
      properties:
        system:
          type: string
          format: uri
        code:
          type: string
        display:
          type: string
    Period:
      type: object
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
...
    Coverage:
      type: object
      required: [resourceType, status, beneficiary]
      properties:
        resourceType:
          type: string
          enum: [Coverage]
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        status:
          type: string
          enum: [active, cancelled, draft, entered-in-error]
        type:
          $ref: '#/components/schemas/CodeableConcept'
        subscriber:
          $ref: '#/components/schemas/Reference'
        beneficiary:
          $ref: '#/components/schemas/Reference'
        payor:
          type: array
          items:
            $ref: '#/components/schemas/Reference'
        period:
          $ref: '#/components/schemas/Period'
    Claim:
      type: object
      required: [resourceType, status, type, use, patient, provider]
      properties:
        resourceType:
          type: string
          enum: [Claim]
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        status:
          type: string
          enum: [active, cancelled, draft, entered-in-error]
        type:
          $ref: '#/components/schemas/CodeableConcept'
        use:
          type: string
          enum: [claim, preauthorization, predetermination]
        patient:
          $ref: '#/components/schemas/Reference'
        provider:
          $ref: '#/components/schemas/Reference'
        created:
          type: string
          format: date-time
        total:
          type: object
          properties:
            value:
              type: number
            currency:
              type: string
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          description: Total number of matching records
        limit:
          type: integer
        offset:
          type: integer
...